Refactors examples page into modules
Breaks down the monolithic examples.js file into smaller, more manageable modules for better organization and maintainability.

This change improves code readability and allows for easier feature development and debugging.
