+++ public/index.html
@@ -6,9 +6,9 @@
         <title>Custom Component Tester</title>
         <meta name="viewport" content="width=device-width, initial-scale=1" />
         <script type="module" src="/src/components/index.js"></script>
-        <script type="module" src="/examples.js"></script>
+        <script type="module" src="/scripts/examples.js"></script>
         <link rel="stylesheet" type="text/css" href="main.css" />
-        <link rel="stylesheet" type="text/css" href="examples.css" />
+        <link rel="stylesheet" type="text/css" href="/styles/examples.css" />
     </head>
     <body class="developer-tools">
         <aside id="sidebar" class="sidebar"></aside>
